passing Deck item to  DeckItem component
After fetch Desk Data on DesklList, when click on a Desk it has to pass  desk data to DeckItem in order to render it.

In order to use params in the title, we need to make navigationOptions a function that returns a configuration object. It might be tempting to try to use this.props inside of navigationOptions, but because it is a static property of the component, this does not refer to an instance of the component and therefore no props are available. Instead, if we make navigationOptions a function then React Navigation will call it with an object containing { navigation, navigationOptions, screenProps } -- in this case, all we care about is navigation, which is the same object that is passed to your screen props as this.props.navigation. You may recall that we can get the params from navigation through navigation.getParam or navigation.state.params